updates due to adding OSX and Windows as platforms
These changes are not essential I would say, but as there is a pixi.lock file
already in the repository, it seems consistent to update this to be syrchronised
with the pixi.toml.

The version of the file changes from 5 to 6 with this commit. (So my
pixi (version 0.39.3) seems to use 6 while 5 was used when the repository was
created.)